Separate out validate logic and types
