+++ packages/perseus/src/widgets/categorizer/categorizer.tsx
@@ -15,15 +15,12 @@ import mediaQueries from "../../styles/media-queries";
 import sharedStyles from "../../styles/shared";
 import Util from "../../util";
 
+import categorizerValidator from "./categorizer-validator";
+
+import type {Rubric, UserInput} from "./categorizer.types";
 import type {PerseusCategorizerWidgetOptions} from "../../perseus-types";
 import type {PerseusStrings} from "../../strings";
 import type {PerseusScore, WidgetExports, WidgetProps} from "../../types";
-
-type UserInput = {
-    values: ReadonlyArray<number>;
-};
-
-export type Rubric = PerseusCategorizerWidgetOptions;
 
 type Props = WidgetProps<RenderProps, Rubric> & {
     values: ReadonlyArray<string>;
@@ -60,28 +57,7 @@ export class Categorizer extends React.Component<Props, State> {
         rubric: Rubric,
         strings: PerseusStrings,
     ): PerseusScore {
-        let completed = true;
-        let allCorrect = true;
-        rubric.values.forEach((value, i) => {
-            if (userInput.values[i] == null) {
-                completed = false;
-            }
-            if (userInput.values[i] !== value) {
-                allCorrect = false;
-            }
-        });
-        if (!completed) {
-            return {
-                type: "invalid",
-                message: strings.invalidSelection,
-            };
-        }
-        return {
-            type: "points",
-            earned: allCorrect ? 1 : 0,
-            total: 1,
-            message: null,
-        };
+        return categorizerValidator(userInput, rubric, strings);
     }
 
     static getUserInputFromProps(props: Props): UserInput {
@@ -106,7 +82,7 @@ export class Categorizer extends React.Component<Props, State> {
     }
 
     simpleValidate: (arg1: Rubric) => PerseusScore = (rubric) => {
-        return Categorizer.validate(
+        return categorizerValidator(
             this.getUserInput(),
             rubric,
             this.context.strings,
