+++ packages/perseus/src/widgets/categorizer/categorizer.types.ts
@@ -1,0 +1,7 @@
+import type {PerseusCategorizerWidgetOptions} from "../../perseus-types";
+
+export type UserInput = {
+    values: ReadonlyArray<number>;
+};
+
+export type Rubric = PerseusCategorizerWidgetOptions;
